+++ R/zample.R
@@ -41,12 +41,11 @@
 #' # setting replace=TRUE will result in an error
 #' set.seed(42); tryCatch(zample(7, size=2), error=wrap_error)
 #'
-#' @md
 #' @export
 zample = function (x, size=length(x), replace = FALSE, prob = NULL)
 {
     # Bail out of sampling from data.frames, use dplyr::sample_X() for that
-    if (inherits(x,"data.frame"))
+    if (inherits(x, "data.frame"))
     {
         stop("zulutils::zample() does not support data.frames")
     }