+++ tests/testthat/test-zmisc.R
@@ -67,12 +67,12 @@ test_that("zeq works", {
 context("zingle")
 test_that("zingle works", {
 
-    # If all equal, return first item
+    # Check that if all equal, return first item
     expect_equal( zingle(rep(10,10)), 10 )
     expect_equal( zingle(rep("a",20)), "a" )
     expect_equal( zingle(as.factor(rep("a",20))), as.factor("a") )
 
-    # If not all equal,
+    # Check that if not all equal, an error is thrown
     expect_error( zingle(1:2) )
 
     # Check that NAs behave as expected with na.rm
@@ -80,8 +80,11 @@ test_that("zingle works", {
     expect_equal( zingle(c(NA,NA,NA, 1,NA), na.rm=TRUE),   1 )
     expect_equal( zingle(c(NA,NA,NA,NA,NA), na.rm=TRUE),   NA )
 
-    # Without na.rm=TRUE, NAs should cause an error
+    # Check that without na.rm=TRUE, NAs cause an error
     expect_error( zingle(c( 1, 1,NA, 1, 1)) )
     expect_error( zingle(c(NA,NA,NA, 1,NA)) )
     expect_error( zingle(c(NA,NA,NA,NA,NA)) )
+
+    # Check that the result never has names
+    expect_named( zingle(c(a=1,b=1)), NULL)
 })